tests/repoinfo: be more robust when listing local branches
As tests uses the local git client, it inherits from the git global
configuration. So if the git global configuration has been changed
to customize the branch listing like 'branch.sort=-committerdate', branch
are sorted in a different way than the default alphabetical order.

Make the test agnostic to the branch sorting algorithm.

Signed-off-by: Cedric CHEDALEUX <cedric.chedaleux@orange.com>